mobile_automation/requirements.txt
# requirements.txt
Appium-Python-Client
pytest
pytest-html
faker
python-dotenv